requirements.txt
streamlit==1.51.0
python-dotenv==1.0.0
requests==2.31.0
agno==2.2.13
serpapi==0.1.5
openai==0.27.0
